naive_search/pyproject.toml
[project]
name = "aai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.0",
    "fastapi[standard]>=0.128.7",
    "google-genai>=1.63.0",
    "huggingface-hub>=1.4.1",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tavily>=1.1.0",
    "transformers>=5.1.0",
]
